Fixed profile picture (recruitment + people/[n]/edit) and tested it in recruitment
Moved help/image functionality into lib/image to reduce bloating
