+++ people/models.py
@@ -51,7 +51,11 @@ class Profile(models.Model):
     registration_year = models.IntegerField(null=True, blank=True)
     planned_graduation = models.IntegerField(null=True, blank=True)
     linkedin_url = models.URLField(null=True, blank=True)
-
+    
+    portrait = models.ImageField(
+            upload_to=UploadToDir('profiles', 'portrait'),
+            blank=True,
+            )
     picture_original = models.ImageField(
             upload_to=UploadToDirUUID('profiles', 'picture_original'),
             blank=True,
