+++ templates/edit_person.html
@@ -12,15 +12,6 @@
 
     <form method="POST" enctype="multipart/form-data">
     {% csrf_token %}
-                            <div class="form-group">
-                            <label for="image">Image of you</label>
-                            <input type="file" id="image" name="image" accept="image/*">
-                            {% if person.image %}
-                                <br>
-                                <div><img src="/media/{{ person.image }}" style="max-width: 100%; max-height: 400px;"></div>
-                            {% endif %}
-                        </div>
-
     {{ form|crispy }}
     <button class="btn btn-primary">Save</button>
     </form>
