+++ templates/view_person.html
@@ -10,9 +10,9 @@
 
     <div>
 
-                                {% if person.picture %}
-                                <div><img src="/media/{{ person.picture }}" style="max-width: 100%; max-height: 400px;"></div>
-                            {% endif %}
+	{% if person.portrait %}
+	<div><img src="/media/{{ person.portrait }}" style="max-width: 100%; max-height: 400px;"></div>
+	{% endif %}
     <br>
     </div>
